requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-google-genai>=0.0.4
python-dotenv
google-generativeai
streamlit
faiss-cpu
tiktoken
pydub
ffmpeg-python
openai-whisper>=20231117
pytube
yt-dlp>=2023.7.6
transformers>=4.30.0
librosa>=0.10.0
numpy>=1.23.5
scipy>=1.10.0

# Removed:
# moviepy - not used for video processing
# pytube - not used for video downloading 
